refactor(agents): collection_uploader.py を責務別 mixin に分割 (#465)
613 行で責務肥大していた collection_uploader.py を責務別 mixin モジュールへ
分割。#465 の follow-up（youtube_auto_uploader.py 分割 #640 と同方針）。

- tracking / workflow-state I/O → _tracking_io.py (TrackingIOMixin)
- 公開日一覧取得 / publishAt 計算 → _published_dates.py (PublishedDatesMixin)
- プレイリスト自動割り当て → _playlist_assignment.py (PlaylistAssignmentMixin)
- Complete Collection 実行ループ → _complete_collection_executor.py
  (CompleteCollectionExecutorMixin)
- 定数 → _collection_uploader_constants.py
- 本体 613 → 355 行に縮小

機能・公開 API・挙動はすべて不変。既存テストが直接 patch する
`patch("youtube_automation.agents.collection_uploader.PlaylistManager")` /
`patch("...load_config")` / `patch("...YouTubeAutoUploader")` のポイントは
本モジュールでの再エクスポートで温存（mixin 側は collection_uploader 経由で
PlaylistManager/load_config を遅延解決して mock を尊重する）。

回帰テストの literal Path ガード (`tests/test_collection_paths.py::_MIGRATED_FILES`)
に新 mixin 4 ファイルを追加して制約を継承。

関連テスト 162 件 (test_collection_uploader / test_collection_paths /
test_youtube_auto_uploader / test_schedule_cadence / test_short_uploader) green。

Closes #465
