+++ src/youtube_automation/agents/_tracking_io.py
@@ -1,0 +1,107 @@
+"""tracking JSON / workflow-state JSON の I/O を提供する mixin。
+
+責務分割（Issue #465）の一環で ``collection_uploader.py`` から分離した。
+挙動は分割前と同一で、``self.config`` / ``self.collections_root`` 等は
+合成先クラス（``CollectionUploader`` 本体）が提供する。
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+from pathlib import Path
+
+from youtube_automation.agents._collection_uploader_constants import (
+    TRACKING_STATUS_COMPLETED,
+    WORKFLOW_PHASE_COMPLETE,
+    WORKFLOW_STAGE_LIVE,
+)
+from youtube_automation.utils.collection_paths import CollectionPaths
+from youtube_automation.utils.schedule import now_in_schedule_tz
+
+logger = logging.getLogger(__name__)
+
+
+class TrackingIOMixin:
+    """tracking JSON / workflow-state JSON の I/O を提供する mixin。"""
+
+    def _get_tracking_path(self, collection_path: Path) -> Path:
+        return CollectionPaths(collection_path).tracking_path
+
+    def _load_tracking(self, collection_path: Path) -> dict | None:
+        """tracking ファイル読み込み"""
+        tracking_file = self._get_tracking_path(collection_path)
+        if not tracking_file.exists():
+            return None
+
+        try:
+            with open(tracking_file, "r", encoding="utf-8") as f:
+                return json.load(f)
+        except Exception:
+            return None
+
+    def _save_tracking(self, collection_path: Path, tracking: dict):
+        """tracking 保存"""
+        tracking_file = self._get_tracking_path(collection_path)
+        tracking_file.parent.mkdir(exist_ok=True)
+        try:
+            with open(tracking_file, "w", encoding="utf-8") as f:
+                json.dump(tracking, f, indent=2, ensure_ascii=False)
+        except Exception as e:
+            logger.warning(f"⚠️  追跡ファイル保存エラー: {e}")
+
+    def _completed_tracking_record(self, complete_video: dict, publish_at: str | None) -> dict:
+        record = {
+            "video_id": complete_video["video_id"],
+            "video_url": complete_video["video_url"],
+            "upload_time": now_in_schedule_tz(self.config).isoformat(),
+            "publish_at": publish_at,
+            "status": TRACKING_STATUS_COMPLETED,
+        }
+        if complete_video.get("upload_source"):
+            record["upload_source"] = complete_video["upload_source"]
+        return record
+
+    def _update_workflow_upload(self, collection_path: Path, complete_video: dict, publish_at: str | None) -> None:
+        ws_path = CollectionPaths(collection_path).workflow_state_path
+        if not ws_path.exists():
+            return
+
+        state = json.loads(ws_path.read_text(encoding="utf-8"))
+        upload = state.get("upload")
+        if not isinstance(upload, dict):
+            raise ValueError(f"workflow-state.json upload must be object: {ws_path}")
+
+        updated_upload = {
+            **upload,
+            "video_id": complete_video["video_id"],
+            "video_url": complete_video["video_url"],
+            "publish_at": publish_at,
+        }
+        updated_state = {
+            **state,
+            "upload": updated_upload,
+            "updated_at": now_in_schedule_tz(self.config).isoformat(),
+        }
+        if collection_path.parent.name == WORKFLOW_STAGE_LIVE:
+            updated_state = {
+                **updated_state,
+                "stage": WORKFLOW_STAGE_LIVE,
+                "phase": WORKFLOW_PHASE_COMPLETE,
+            }
+        ws_path.write_text(json.dumps(updated_state, indent=2, ensure_ascii=False), encoding="utf-8")
+
+    def _initialize_tracking(self, collection_path: Path) -> dict:
+        """tracking を初期化"""
+        tracking = {
+            "schema_version": 3,
+            "collection_name": collection_path.name,
+            "status": "in_progress",
+            "complete_collection": {"status": "pending"},
+        }
+
+        self._save_tracking(collection_path, tracking)
+        return tracking
+
+
+__all__ = ["TrackingIOMixin"]
